Fix c3twin interview
Page: "HBQ School and Peer" -- we want to get rid of the "HBQ2-A"
question and just start with "HBQ2-1"
Page: "HBQ1 School and Peer Cotwin" -- (SAME) we want to get rid of the
"HBQ1-A" question and just start with "HBQ1-1"
Page: "Sensory Defensiveness" -- ALL of these questions are not
supposed to be able to enter a "2" but they all have that option. They
should only be able to enter values "0, 1, 9999, and 9998"
Page: "Sensory Defensiveness" -- Question "31.Other ways you are touch
sensitive" AND Question "23.Other ways you are sensitive to touch" need
comment boxes where you can enter phrases/words- not number values like
they have currently.
Page: "APES" -- Question "94" (the comment box) should NOT be called
"94" because it is actually associated with 93. It should only be
called "Explain:"
The entire Height Weight Measures Page needs to be deleted (there are
no questions in the Twin interview that ask these questions...they are
only part of the Care Giver interview